Classess Cache and Request

+++ src/hashcode2017/Cache.java
@@ -26,7 +26,7 @@ public class Cache {
 		return size;
 	}
 
-	public void setSize(int size) {
+	public void setSize(double size) {
 		this.size = size;
 	}
